create result storage structure
